use get_slice to copy lists
